requirements.txt
django==5.1.3
psycopg2-binary==2.9.10
djangorestframework==3.15.2
drf-yasg==1.21.8
